backend/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
drf-spectacular>=0.27,<1.0
mysqlclient>=2.2,<3.0
